fix(notifications): keep sync review dialog on current route
Move background sync review handling into a global renderer host so notification activations can open the review dialog without leaving the current screen.

Forward the desktop notification shortcut through Electron to keep Cmd+Shift+I working when task detail terminals own keyboard focus.
