+++ src/desktop/renderer/components/BoardCommandProvider.tsx
@@ -189,6 +189,20 @@ export function BoardCommandProvider({ children }: PropsWithChildren) {
     };
   }, [isTaskQuickSearchOpen]);
 
+  useEffect(() => {
+    const unsubscribe = window.kanvibeDesktop?.onNotificationShortcut?.(() => {
+      if (isTaskQuickSearchOpen) {
+        return;
+      }
+
+      notificationCenterHandlerRef.current?.();
+    });
+
+    return () => {
+      unsubscribe?.();
+    };
+  }, [isTaskQuickSearchOpen]);
+
   const value = useMemo<BoardCommandContextValue>(() => ({
     canCreateBranchTodo,
     registerBoardHandlers,
