+++ src/types/kanvibeDesktop.d.ts
@@ -10,8 +10,9 @@ interface KanvibeDesktopApi {
   activateNotification?: (notificationId: string) => Promise<boolean>;
   consumePendingNotificationActivation?: () => Promise<AppNotification | null>;
   onNotificationActivated?: (listener: (notification: AppNotification) => void) => () => void;
+  onNotificationShortcut?: (listener: () => void) => () => void;
   onCreateTaskShortcut?: (listener: () => void) => () => void;
-  [key: string]: any;
+  [key: string]: unknown;
 }
 
 interface Window {
